+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.23.0] - 2020-06-01
+
 ### Added
 - Improved indexing for FieldtypeOptions.
 - getSQL() method and sql property for the Query class.